chore: static analysis
Updated phpunit.xml to use the correct xsi:noNamespaceSchemaLocation value. Updated the ApplicationObserver to include a default case in the match statement. Renamed the laravel workflow file to tests.yml. Updated the UpdateRefreshTokenAction to use updateOrCreate. Updated the AuthenticationServiceProvider to use app() instead of $this->app. Removed unnecessary comments in the UserFactory and SsoController files. Added type hints to functions in the CharacterUser model. Updated the BuildUserLevelRequiredScopes service to handle the 'global_scope' attribute. Added type hints to functions in the Role, Affiliation, and AclMember models.